dockerfile, docker-compose, nginx

+++ Jenkinsfile
@@ -9,18 +9,18 @@
            stage('Build and Deploy') {
                steps {
                    script {
-                       //sh 'docker rm -f $(docker ps -aq)' // Удаляем старые контейнеры
-                       script {
                     // Получаем список всех контейнеров
-                            def containers = sh(script: 'docker ps -aq', returnStdout: true).trim()
+                    def containers = sh(script: 'docker ps -aq', returnStdout: true).trim().split("\s+")
                     
                     // Проверяем, есть ли контейнеры для удаления
-                            for (container in containers) {
-                                sh "docker rm -f ${container}"
-                            }
-                       sh 'docker-compose up -d'
-                        } // Собираем и запускаем новые контейнеры
-                   }
+                    if (containers.size() > 0 && containers[0] != '') {
+                        // Удаляем каждый контейнер по отдельности
+                        for (container in containers) {
+                            sh "docker rm -f ${container}"
+                        }
+                    } else {
+                        echo 'No containers to remove.'
+                    }
                }
            }
        }
